requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.5.0
anyio==3.7.1
array-record==0.4.1
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
backoff==2.2.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cohere==4.27
cryptography==41.0.4
dataclasses-json==0.6.1
dm-tree==0.1.8
etils==1.5.0
fastavro==1.8.2
Flask==2.3.3
Flask-Cors==4.0.0
flask-restx==1.1.0
frozenlist==1.4.0
fsspec==2023.9.2
googleapis-common-protos==1.60.0
greenlet==3.0.0
grpcio==1.59.0
grpcio-tools==1.59.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
langchain==0.0.317
langsmith==0.0.46
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.7
numpy==1.26.0
packaging==23.2
portalocker==2.8.2
promise==2.3
protobuf==4.24.3
psutil==5.9.5
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pypdf==3.16.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qdrant-client==1.5.4
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
tenacity==8.2.3
tensorflow-metadata==1.13.1
termcolor==2.3.0
tfds-nightly==4.9.3.dev202310040044
toml==0.10.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==1.26.17
validators==0.22.0
weaviate-client==3.24.2
Werkzeug==2.3.7
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0
